Adding the transaction form and more
- Authentication and protected
Some pages it will automaticly redirect you if a certain condition matches. This is to enhance the UX and create less confusion

- Transactions front-end

Added a form to add transactions and a select box with options to choose from. select box is also dynamic so a lot of options to choose from