Fixed an issue w/ Dashboard
Fixed an issue with the dashboard page where if you reload it you get an error.

Kinda a hacky way. But it makes the page more faster and it creates more ways to avoid the render limit of react

Also regarding the database.sql. It was to improve the charts and statistics for the player to get a better insight instead of the old way