Restructiring
A restruction of the infrastructure of the web application.

I noticed that I have done a lot of mistakes in this particular project. In order to save me a lot of work in the future I decided to restructure a part of the web application. This will be the start.